Aadi-Voice-Assistant/requirements.txt
speechrecognition
pyttsx3
customtkinter
groq
python-dotenv
psutil
numpy